Fixed parsing of conditional port expression.
The *conditional port expression* (`cond_expr`) is part of the `COND`
conditional path delay. Past implementation did not handle bracketing.
New implementation split the grammar between binary expressions and
unary expressions.

Presently we do care only for the expression string and not about
(binary) operator precedence. Hence we do not build the (precedence)
binary tree for the expression. The present implementation is likely
to have better performance and if the operator precedence determination
were needed, I would suggest to have a separate, on-demand parser for
it.

Along with the parser change, minor changes/fixes were made to the lexer.
Namely the `ARITHMETIC` token was broken into individual symbols,
`PLUS`, `MINUS` and `TIMES`. This was due to `PLUS` and `MINUS` acting
as both unary and binary operators.

`SCALARCONSTANT` was fixed not to match strings `0'b1` and `0'b0`, which
are not valid constants.