Uniquification of names of timing annotations.
When adding collected timing annotations under the cell construct,
it may happen that there will be two or more annotations with the
same name. For example, two conditional IOPATH delays (with different
conditions) for the same from-to pair of pins. The way how the
derived timing annotations are named and indexed under the cell
will yield name collision.

This commit adds code to detect such collisions and append a sequential
number to uniquify every colliding occurrence. For example::

    (CELL
        (CELLTYPE "AND2")
        (INSTANCE top.b.d)
        (DELAY
            (ABSOLUTE
                (COND !en (IOPATH a y (1:2:3)(1:2:3)))
                (COND en  (IOPATH a y (4:5:6)(4:5:6)))
            )
        )
    )

The first annotation will be named `iopath_a_y` and the second
`iopath_a_y_1`.