Changed `RETAIN` organization in parsed structure.
Originally, the `RETAIN` delay value was added as a separate `iopathretain`
timing arc/record in the parsed structure of a cell. This turned to be bad
idea for recovering the SDF structure.

Newly, the retain delay is put under its parent `iopath`. The retain delval
list goes under the `retain_paths` dictionary key. This key is populated only
when the `IOPATH` has an associated `RETAIN`; otherwise the `retain_paths`
key will not exist. The existince of `RETAIN` can be tested through a bool
key `has_retain` that will be populated for every `iopath` type timing
arc/record structure.