Enriched options and functions of CLI utilities.
Added support for gzip'ed files (at input and output). Added option to write
output to `stdout` and changed utils messaging to go to `stderr`. Added `--force`
option to overwrite existing files.

For `sdf2json.py` changed `--dir` to be `None` by default so that the converted
JSON is created next to the original input file. This will make it more flexible
when giving multiple files at input, each of which resides elsewhere.